Pluggable infrastructure + modules
Includes changes to create a pluggable infrastructure for FireFly UI
and introduces the concepts of `modules` and `core`. Modules are self-contained
applications with their own translations, routing, navigation, and components
that live within FireFly UI. FireFly UI `core` is a set of common components,
hooks, utils, and contexts that can be reused across modules.

Navigation has been reworked to allow the user to navigate between different modules.

The original FireFly Explorer is now the `data` module. Two other modules,
`tokens` and `network-map`, are being introduced in this PR.

Signed-off-by: Alex Shorsher <alex.shorsher@kaleido.io>