Only replace "Transaction Submitted" with transaction name on split views
When viewing detailed lists of mixed types of events, it's more helpful to see the
event type of "Transaction Submitted" vs. the type of transaction that was submitted.

Only fill in the specific transaction name when looking at a filtered list that
contains nothing but "Transaction Submitted" events (since it provides valuable
extra info in that context).

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>